code refactored ,added env , removed major bugs
