+++ apps/web/components/Landing/navbar/navbar.css
@@ -1,0 +1,39 @@
+.open{
+    position: absolute;
+    top:90px;
+    left: 0;
+    width: 100%;
+    height: fit-content;
+    margin-left: 2px;
+    margin-right: 2px;
+    display: flex;
+    justify-content: space-around;
+    margin: 0;
+}
+.open a {
+    padding-left: 1px;
+    padding-right: 1px;
+}
+
+  @media screen and (min-width: 768px) and (max-width: 1098px) {
+    .nav_flex{
+      width:unset;
+    }
+  }
+  @media screen and (min-width: 768px) and (max-width: 920px) {
+    .nav_flex{
+      width:unset;
+    }
+    .anchor_class{
+      font-size: 1rem;
+      line-height: 1rem;
+      letter-spacing: 0px;
+      padding-left: 0px;
+      padding-right: 0px;
+    }
+    .anchor_links{
+      font-size: 1rem;
+      padding-left: 1.25rem;
+      padding-right: 1.25rem;
+    }
+  }
