+++ apps/web/components/Teams/Animation/TeamAnimation.tsx
@@ -34,7 +34,7 @@ const TeamAnimation = () => {
   };
 
   const getData = () => {
-    fetch("https://us-central1-techspardha-87928.cloudfunctions.net/api2/contacts")
+    fetch(`${process.env.NEXT_PUBLIC_BASE_URL}/contacts`)
       .then(res => res.json())
       .then((res: ContactsResponse) => {
         const array: Members[] = [];
